refactor: Extract MEV transaction table into separate component
- Create MEVTransactionTable component with transactions display logic
- Move helper functions (normalizeMevType, getMevTypeColor, hexToRgba, formatTxHash) to new component
- Extract transaction-related CSS to MEVTransactionTable.css with mev-tx-table- prefix
- Simplify LiveSection by replacing inline table with MEVTransactionTable component
- Remove unused imports from LiveSection (Table, TableBody, TableCell, etc.)
- Update responsive styles to remove reference to removed transaction card class
